cart_view, order maker, update, delete items from cart

+++ src/main/java/com/service/UserService.java
@@ -38,7 +38,7 @@ public class UserService {
         userClient.saveUser(user);
     }
 
-    public void delete(Integer userId) {
+    public void delete(Long userId) {
         userClient.deleteUser(userId);
     }
